% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addModels.R
\name{addModels}
\alias{addModels}
\title{addModels}
\usage{
addModels(screen, alternative, ROPE, c = 0)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column}

\item{alternative}{String indicating direction of comparison. Must be either
'less' or 'greater'}

\item{ROPE}{Region of Practical Equivalence a practical cutoff at which a
compound could be active but so minimally so we do not care}

\item{c}{Numeric value for the inactive popluation in RVM t-test. Use same
value as ROPE.
values}
}
\description{
Add z-score, RVM t-test, Bayesian RVM, and custom Bayesian models to HTS
screening data
}
\examples{
Load screen
data(mdmx)
mdmx <- applyModels(mdmx)
}

