% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rope.R
\name{autoRope}
\alias{autoRope}
\title{autoRope}
\usage{
autoRope(assay.conc, max.conc, modifier = 1, hillslope = 1, top = 100,
  bottom = 0)
}
\arguments{
\item{assay.conc}{numeric value of assay concentration (same units as max.con)}

\item{max.conc}{numeric value of maximum detectable EC50 (same units as assay.conc)}

\item{modifier}{numeric modifier of signal}
}
\description{
From a provided maximum detectable EC50 and by assuming a one site effective
concentration 50 curve this method estimates the minimum signal cutoff for an
active hit at the given screening concentration.
}
\examples{
autoRope(25, 100, 0.8)
}

