% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkAlternative}
\alias{checkAlternative}
\title{checkAlternative}
\usage{
checkAlternative(alternative)
}
\arguments{
\item{alternative}{Character string specifying the alternative hypothesis
must be one of: 'less', 'greater'}
}
\description{
Check alternative flag is one of 'less' or 'greater'
}
\examples{
checkAlternative('less')
}

