% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checks.R
\name{checkRole}
\alias{checkRole}
\title{checkRole}
\usage{
checkRole(role)
}
\arguments{
\item{role}{Character string specifying well role must be one of: 'less',
'greater'}
}
\description{
Check role flag is one of 'positive control', 'neutral control', 'compound'
}
\examples{
Load screen
checkRole('positive control')
}

