% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checks.R
\name{checkScreen}
\alias{checkScreen}
\title{checkScreen}
\usage{
checkScreen(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID,
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
Check screen data for necessary data labels and values
}
\examples{
Load screen
data(mdmx)
checkScreen(mdmx)
}

