% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{columnPlot}
\alias{columnPlot}
\title{columnPlot}
\usage{
columnPlot(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
Make a column-wise plot of a high-throughput screen
}
\examples{
Load screen
data(mdmx)
columnPlot(mdmx)
}

