% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/curveFit.R
\name{curveFit}
\alias{curveFit}
\title{curveFit}
\usage{
curveFit(x, y, top.mu, bottom.mu, runjags.method = "rjparallel")
}
\arguments{
\item{x}{Numeric of compound concentrations}

\item{y}{Numeric measured values aligned with x}
}
\description{
When the only tool you have is a hammer, every problem looks like a nail.
curveFit, fits a 4 parameter logistic inhibition curve according to
y[i] ~ N(mu[i], tau)
mu <- top + (bottom - top)/(1 + 10^((log(IC50) - log(x))*hillslope))
}
\details{
IC50 <- exp(1/100)
hillslope ~ dunif(0.1, 3)
top ~ dnorm(0, 0.01)
bottom ~ dnorm(-100, 0.01)
tau <- pow(sigma, -2)
sigma ~ dunif(0, 100)
}
\examples{
x <- c(2, 4, 8, 16, 32, 64, 128, 256)
y <- -100/(1 + 10^((log(27) - log(x))*1.23))
y.error <- y + rnorm(length(y), 0, 5)
plot(y.error, x)
estimate <- curveFit(x, y.error)
estimate
}

