% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAB.R
\name{flik}
\alias{flik}
\title{flik}
\usage{
flik(param, variance, m)
}
\arguments{
\item{param}{A list containing the values for a and b.}

\item{variance}{A vector of the observed variances}

\item{m}{A vector of degrees of fredom for each observed variance 
(number of observations - 1)}
}
\description{
This function calculates the log likelihood that the ratio of 
variance and the expected value of the fitted distribution come from an F
distribution with m and 2*a degrees of freedom. Where m is number of 
observations - 1, and a is the alpha parameter for the inverse gamma 
distribution.
}

