% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReps.R
\name{getReps}
\alias{getReps}
\title{getReps}
\usage{
getReps(screen, role, variable)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column
values}

\item{role}{character string indicating which well type to subset data must
be one of: 'positive control', 'negative control', or 'compound'}

\item{factor}{column name containing factor by which to cast data}
}
\description{
Get replicates by factor
}
\examples{
Load screen
data(mdmx)
compound.replicates <- getReps(mdmx, 'compound', 'Compound_ID')

neutral.replicates <- getReps(mdmx, 'neutral control', '')
}

