% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{logisticFit}
\alias{logisticFit}
\title{logisticFit}
\usage{
logisticFit(x, top, bottom, IC50, hillslope)
}
\arguments{
\item{x}{concentration}

\item{IC50}{log(IC50)}
}
\description{
4 parameter logistic binding function
}
\examples{
Load screen
data(mdmx)
wellHeatMap(mdmx)
}

