% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{nBayes}
\alias{nBayes}
\title{nBayes}
\usage{
nBayes(screen, alternative, ROPE, mu0 = 0, n0 = 0.01, size = 20000)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
variables: Plate_ID, Row, Column, Compound_ID, Well_Type, Plate_Mask, 
Well_Mask, Value. For an explination of variables see read me.}

\item{alternative}{A character string specifying the alternet hypothesis. 
Must be one of "two.sided", "greater", or "less".}
}
\description{
Deprecated: Estimate the probability of a compound coming from the neutral
control distribution assuming the compounds and neutral controls are normally 
distributed with unknown mu and tau.
}
\examples{
Load screen
data(mdmx)
mdmx <- nBayes(mdmx, alternative='less')
head(mdmx)
}

