% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{nContrast}
\alias{nContrast}
\title{nContrast}
\usage{
nContrast(x, a, b, mu0, n0, null.post, alternative, size, ROPE)
}
\arguments{
\item{x}{an array of data}

\item{a}{alpha paramter}

\item{b}{beta parameter}

\item{mu0}{mu ~ N(mu0, n0)}

\item{n0}{mu ~ N(mu0, n0)}

\item{null.post}{data.frame or data.table of a null posterior with column mu}
}
\description{
Depricated: estimate the probability a compound is from the neutral
distribution
}
\examples{
set.seed(7)
compound <- rnorm(2, 100, 10)
null.reference <- rnorm(32, 0, 10)
a <- 30
b <- 3100
mu0 <- 0.0
n0 <- 0.0001
size <- 20000

null.post <- nSample(null.reference, a, b, mu0, n0, size)
posterior <- nContrast(compound, a, b, mu0, n0, null.post, 
                       alternative='greater', size)
posterior
}

