% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample.R
\name{nSample}
\alias{nSample}
\title{nSample}
\usage{
nSample(x, a, b, mu0, n0, size)
}
\arguments{
\item{x}{an array of data}

\item{a}{alpha paramter}

\item{b}{beta parameter}

\item{mu0}{mu ~ N(mu0, n0)}

\item{n0}{mu ~ N(mu0, n0)}
}
\description{
Depricated: sample a posterior distribution of the form
  mu|x,tau ~ N(mu, tau)
where mu and tau are unknown
}
\examples{
compound <- rnorm(2, 100, 10)
a <- 30
b <- 3100
mu0 <- 0.0
n0 <- 0.0001
size <- 20000

posterior <- nSample(compound, a, b, mu0, n0, size)
hist(posterior$mu)
}

