% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{percentEffect}
\alias{percentEffect}
\title{percentEffect}
\usage{
percentEffect(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID,
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
Calculate percent effect relative to positive and neutral controls
}
\examples{
Load screen
data(mdmx)
mdmx <- percentEffect(mdmx)
}

