% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platePattern.R
\name{platePattern}
\alias{platePattern}
\title{platePattern}
\usage{
platePattern(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
Create a heat map of plate wells by run
}
\examples{
library(RColorBrewer)
library(ggplot)
Load screen
data(mdmx)
pattern <- platePattern(mdmx)
myPalette <- colorRampPalette(rev(brewer.pal(11, "Spectral")))
ggplot(pattern, aes(factor(Column), factor(Row))) +
 facet_wrap('Run_ID') + 
 geom_tile(aes(fill=Background), color='white') +
 scale_fill_gradientn(colours = myPalette(100), limits=c(-150, 100)) +
 ggtitle('Plate Well Heatmap') + 
 xlab('Column') + 
 ylab('Row') +
 theme(axis.text=element_text(size=16), 
       axis.title=element_text(size=16), 
       legend.text=element_text(size=16),
       title=element_text(size=16),
       legend.title=element_blank())
}

