% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{plateStats}
\alias{plateStats}
\title{plateStats}
\usage{
plateStats(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID,
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
Add number of observations, plate median, and plate MAD
}
\examples{
Load screen
data(mdmx)
mdmx <- plateStats(mdmx)
head(mdmx)
}

