% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotCurve}
\alias{plotCurve}
\title{plotCurve}
\usage{
plotCurve(screen)
}
\arguments{
\item{compound}{data.frame row with Compound_ID, IC50, IC50_Error, Hillslope,
Top, Bottom, Value, and Concentration.}
}
\description{
Plot fitted binding curves
}
\examples{
Load screen
data(mdmx)
wellHeatMap(mdmx)
}

