% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plotCurves}
\alias{plotCurves}
\title{plotCurves}
\usage{
plotCurves(compound)
}
\arguments{
\item{compound}{data.frame row with Compound_ID, IC50, IC50_Error, Hillslope,
Top, Bottom, Value, and Concentration.}
}
\description{
Plot fitted binding curves
}
\examples{
Load screen
data(mdmx)
wellHeatMap(mdmx)
}

