% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReps.R
\name{repScatter}
\alias{repScatter}
\title{repScatter}
\usage{
repScatter(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
A function to tie together extracting compound and control replicates for 
plotting
}
\examples{
Load screen
data(mdmx)
replicates <- repScatter(mdmx)
}

