% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{RVMtTest}
\alias{RVMtTest}
\title{RVMtTest}
\usage{
RVMtTest(screen, alternative, c = 0)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
This function calcluates the compounds t-statistic from a Random Variance
Model t-test
}
\details{
References:
Malo, N. et al. Experimental design and statistical methods for improved hit 
detection in high-throughput screening. J. Biomol. Screen. 15, 990–1000 
(2010).
}
\examples{
Load screen
data(mdmx)
mdmx <- RVMtTest(mdmx)
head(mdmx)
}

