% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizations.R
\name{scatterPlot}
\alias{scatterPlot}
\title{scatterPlot}
\usage{
scatterPlot(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID,
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
Make a replicate scatter plot of a primary high-throughput screen
}
\examples{
Load screen
data(mdmx)
scatterPlot(mdmx)
}

