% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFit.R
\name{screenFit}
\alias{screenFit}
\title{Take HTS confirmation in dose and fit curves using curveFit}
\usage{
screenFit(screen)
}
\arguments{
\item{screen}{data.frame or data.table of confirmation in dose screen}
}
\description{
Take HTS confirmation in dose and fit curves using curveFit
}

