% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{screenPlot}
\alias{screenPlot}
\title{screenPlot}
\usage{
screenPlot(screen)
}
\arguments{
\item{screen}{data.frame row with Compound_ID, Value, and Concentration.}
}
\description{
Plot fitted binding curves for a concentration-response screen
}
\examples{
Load screen
data(mdmx)
screenPlot(confirmation)
}

