% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{tBayes}
\alias{tBayes}
\title{tBayes}
\usage{
tBayes(screen, alternative, ROPE)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
variables: Plate_ID, Row, Column, Compound_ID, Well_Type, Plate_Mask, 
Well_Mask, Value. For an explination of variables see read me.}

\item{alternative}{Character string specifying the alternet hypothesis. 
Must be one of "two.sided", "greater", or "less".}

\item{runjags.method}{Character string specifying the method by which jags is 
run, default is set to 'simple' to avoid conflict with snow library}
}
\description{
Estimate the probability of a compound coming from the neutral control 
distribution assuming the compounds and neutral controls are t distributed 
with unknown mean, and variance. The compound means are assumed to come from
t-distribution fit to all compounds in the screen.
}
\examples{
Load screen
data(mdmx)
mdmx <- tBayes(mdmx, alternative='less')
head(mdmx)
}

