% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{tContrast}
\alias{tContrast}
\title{tContrast}
\usage{
tContrast(data, mu.prior, var.prior, nu.prior, a, b, null.post, ROPE,
  alternative, runjags.method = "rjparallel")
}
\arguments{
\item{data}{activity data for a compound}

\item{mu.prior}{location parameter for prior distribution of mu, typically the average of 
all compound activity.}

\item{var.prior}{spread parameter for prior distribution of mu, typically the square 
standard error of the mean for all compounds.}

\item{a}{alpha parameter for prior distribution of compound variance. See EstimateAB.}

\item{b}{beta parameter for prior distribution of compound variance. See EstimateAB.}

\item{null.post}{a MCMC sampling of a null posterior with column name mu.}

\item{alternative}{a character string specifying the relationship to the null posterior, 
must be one of "two.sided", "greater", or "less".}

\item{runjags.method}{method by which jags is run, default is rjparallel}

\item{rope}{Region of Practical Equivalence. This value specifies how close 
the sampled posterior needs to be to the null posterior to consider them 
equivalent, default value is 0.}
}
\description{
Wraps the tSample function and compares its markov chain to a provided null
markov chain. Returns estimated mean, standard deviation, and probability of 
belonging to the null posterior.
}
\examples{
compound <- rnorm(2, 100, 10)
null.reference <- rnorm(320, 0, 10)
pooled <- c(compound, null.reference)
mu.prior <- median(pooled)
var.prior <- var(pooled)
nu.prior <- 1
a <- 30
b <- 3100
ROPE <- 15
null.post <- tSample(null.reference, mu.prior, var.prior, nu.prior, a, b)
posterior <- tContrast(compound, mu.prior, var.prior, nu.prior, a, b, 
                       null.post, ROPE, alternative='greater')
posterior
}

