% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample.R
\name{tParallel}
\alias{tParallel}
\title{tParallel}
\usage{
tParallel(data, mu.prior, var.prior, nu.prior, a, b, null.post, ROPE,
  alternative, runjags.method = "simple")
}
\arguments{
\item{data}{compound activity data with individual compounds in rows and
replicates in columns}

\item{mu.prior}{location parameter for prior distribution of mu. Typically the average of
all compound activity.}

\item{var.prior}{spread parameter for prior distribution of mu. Typically the square
standard error of the mean for all compounds.}

\item{a}{alpha parameter for prior distribution of compound variance. See EstimateAB.}

\item{b}{beta parameter for prior distribution of compound variance. See EstimateAB.}

\item{null.post}{a MCMC sampling of a null posterior with column name mu.}

\item{alternative}{a character string specifying the relationship to the null posterior,
must be one of "two.sided", "greater", or "less".}

\item{runjags.method}{method by which jags is run, default is set to 'simple' to
avoid conflict with snow library}

\item{rope}{Region of Practical Equivalence. This value specifies how close
the sampled posterior needs to be to the null posterior to consider them
equivalent, default value is 0.}
}
\description{
This function runs tContrast sampling in parallel
}
\examples{
set.seed(7)
compounds <- matrix(rnorm(20, seq(0, 100, 10), 3), nrow=10)
null.reference <- rnorm(32, 0, 3)
pooled <- c(null.reference, sapply(compounds, function(x) x))
mu.prior <- median(pooled)
var.prior <- var(null.reference)
nu.prior <- 1
a <- 30
b <- 3100
ROPE <- 15

null.post <- tSample(null.reference, mu.prior, var.prior, nu.prior, a, b)
posteriors <- tParallel(compounds, mu.prior, var.prior, nu.prior, a, b,
                        null.post, ROPE, alternative='greater')
rbindlist(posteriors)
compounds
}

