% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{tSample}
\alias{tSample}
\title{tSample}
\usage{
tSample(data, mu.prior, var.prior, nu.prior, a, b, lambda = 1/29,
  runjags.method = "rjparallel")
}
\arguments{
\item{data}{list of compound activity data}

\item{mu.prior}{location parameter for prior distribution of mu, typically the average of 
all compound activity.}

\item{var.prior}{spread parameter for prior distribution of mu, typically the square 
standard error of the mean for all compounds.}

\item{a}{alpha parameter for prior distribution of compound variance. See EstimateAB.}

\item{b}{beta parameter for prior distribution of compound variance. See EstimateAB.}

\item{lambda}{paramater for normality of t-distribution. Defaults to 1/29. 
Based on the Bayesian Estimation Superseeds T-test http://www.indiana.edu/~kruschke/BEST/BEST.pdf.}

\item{runjags.method}{method by which jags is run, default is rjparallel}
}
\description{
Samples the posterior distribution of a compound's activity according to: \cr
\cr
data ~ dt(mu, tau, nu) \cr
mu ~ dt(mu.prior, tau.prior, nu.prior) \cr
tau ~ dgamma(a, b) \cr
nu ~ dexp(lambda) \cr 
\cr
The sampling is done by MCMC via autorun.jags.
}
\examples{
compound <- rnorm(2, 100, 10)
mu.prior <- 0
var.prior <- 100
nu.prior <- 1
a <- 30
b <- 3100
posterior <- tSample(compound, mu.prior, var.prior, nu.prior, a, b)
hist(posterior$mu, 100)
hist(posterior$tau, 100)
}

