% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wellPlatePair}
\alias{wellPlatePair}
\title{wellPlatePair}
\usage{
wellPlatePair(screen)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
Concatenate Compound_Plate and Well coordinates to create new Well_Plate_Pair
column
}
\examples{
Load screen
data(mdmx)
mdmx <- wellPlatePair(mdmx)
}

