% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{zScore}
\alias{zScore}
\title{zScore}
\usage{
zScore(screen, alternative)
}
\arguments{
\item{screen}{A data frame or data table containing screening data with 
columns: Plate_ID, Row, Column, Compound_ID, Well_Type, Run_ID, 
Compound_Plate, Value. See documentation for further description of column
values}
}
\description{
This function calcluates the compounds z-score
}
\examples{
Load screen
data(mdmx)
mdmx <- zScore(mdmx)
head(mdmx)
}

